requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
openai==1.12.0
pypdf2==3.0.1
python-dotenv==1.0.1
markdown2==2.4.12
pydantic==2.6.3
jinja2>=3.1.2
reportlab==4.1.0
